Add a strong CSP
This prevents all JS from running, and disallows any CSS and images
apart from that loaded from `self`. This is nice from a security point
of view, and since there's really no need for fancy stuff on the client
in such a simple app, may as well go for it :padlock:

I've had to inline the govuk_template layout to remove some inline JS.
This means I can also get rid of the fonts properly though, which is
nice.