Add tests for controllers
Also updated all the services to return null if their env vars aren't
set (instead of erroring). Since I don't know how to mock the services
out it's either this or have them actually send emails in the tests
(which is definitely a bad idea).

These tests should be good enough for now to prevent me from breaking
the code every time I update it, but it could still do with some unit
tests for the services probably.