Test UserController more thorougly
When I first wrote this project I couldn't be bothered to test anything
that had any side effects (like the GitHub and GOV.UK Notify API calls).

Instead, I made the services that talk to these things silently fail if
they weren't configured. This made it look like the tests were working,
but actually they were testing an error path that just happened to look
a bit like a success, because of shitty error handling.

This shitty error handling made it all the way out into the user
interface, resulting in the increasingly well known:

Success: your reference number is `error-creating-pull-request`

page.

I'd like to start refactoring this project and fixing the crappy error
handling, but I'm not comfortable making changes without proper test
coverage.

This commit adds [webmock](https://github.com/bblimke/webmock) and uses
it to mock out requests to github. This allows us to be much more
thorough about our assertions, including testing that:

- the controller updated the contents of the two terraform files in the
  correct way (to add the user)
- the controller redirected to the confirmation page
- the pull request URL we get back from github is stored in session

I had to do a couple of gross things to get this to work:

- in `read_from_session` we're using
  [rack_session_access](https://github.com/railsware/rack_session_access),
  to get the session. This basically renders an encoded version of the
  session in a `pre` tag, which we then have to fish out and decode.
  This is basically what rack_session_access does in its [capybara helper](https://github.com/railsware/rack_session_access/blob/01675f3eb622f539861904dc01a395d51721b429/lib/rack_session_access/capybara.rb#L13-L18).
- in `assert_content_updated` I had to do some slightly unpleasent
  closure things to capture the body of the request so I could make
  assertions about the content.

If we're happy in general with this approach, it can be extended to the
other controllers. At which point we should be able to do some more
confident refactoring.