Allow sign in without Google in development
Most of this app needs an authenticated user, but it's not trivial to
set up Google SSO locally. This PR adds a magic URL in development that
allows developers to quickly sign in as a test user. Because we only
use Google SSO for login, this enables devs to test most of the apps
functionality.