Gemfile etc.: roll (the 3 different mechanisms we use for determining ruby version) back to 2.7.1
because our default ruby buildpack isn't ready for 2.7.2 yet. instead use
the vagueness operator (`~>`) to allow the github action to choose a newer
ruby if it insists.

this will also require a change in the fourth mechanism we use to choose
ruby version, but that's defined in a whole separate repo