Fix some issues with error handling on the check your answers page
Because Rails, there's inconsistencies with how loading things works
in different environments (autoloading vs eagerloading). I think this
was leading to problems with loading the errors.rb file, at least when
running the tests.

To address this, this commit moves the things in that file in to a
module named like Rails would expect from the filename. This should
mean that Rails finds the right file when attempting to autoload
things.

This commit also changes a few other things that seemed to error, and
adds a test for the case where there's insufficient session data when
the post request is made.