Commit a blank template credentials file
the corresponding master key is included. This is safe to do because the default
credentials file does not contain any secrets or data. It's just a template.

Modify the gitignore file to reflect this change. This follows the pattern described at:

https://discuss.rubyonrails.org/t/rails-master-key-and-per-environment-init/82615/2

Production credentials will be added to a production credentials file and the
corresponding private key will be kept separate and passed to the app as an
environment variable at runtime (using RAILS_PRODUCTION_KEY).

The SECRET_KEY_BASE_DUMMY=1 is documented in the asset pipeline docs and
instructs rails to use the same behaviour as in development where a random
secret_key_base is auto-generated on startup:

https://guides.rubyonrails.org/asset_pipeline.html#local-precompilation