ENG-462: Allow GPA email addresses to log in and request new users and accounts
Folks in the Government Property Agency need to be able to access the
request-an-aws-account app to request new accounts and users and
password resets. Although they are part of the cabinetoffice.gov.uk
Google Workspace, they all use emails with the gpa.gov.uk domain.

Allow gpa.gov.uk emails both basic and advanced access to the app to
allow them to do what they need to do without having to ask Cabinet
Office Digital to do it for them.