Update the cost centres intranet page URL
The location of the cost centres intranet page has changed again. The
old URL is currently redirecting to the new one, but that might stop at
some point, so we need to update the app to use the correct URL. It is
used both in the README and on the organisation page of the app.